+++ binary-codec/types/account_id.go
@@ -35,7 +35,7 @@ func (a *AccountID) FromJSON(value any) ([]byte, error) {
 		return accountID, nil
 
 	case strings.HasPrefix(strValue, "X"):
-		accountID, _, _, err := addresscodec.DecodeXAddress(strValue)
+		accountID, _, _, _, err := addresscodec.DecodeXAddress(strValue)
 		if err != nil {
 			return nil, err
 		}